rename parsers and scripts
