+++ lab_4/plati_digiseller_parser/src/parsers/plati_market.py
